requirements.txt
requests==2.33.1
pandas==3.0.2
numpy==2.4.4
matplotlib==3.10.9
seaborn==0.13.2
scipy==1.17.1
scikit-learn==1.8.0
scikit-posthocs==0.12.0
shap==0.51.0
jupyter==1.1.1
nbconvert==7.17.1
ipykernel==7.2.0
pillow==12.2.0
